feat(hiddentext-detect): add script for opaque text and text hidden under image
